+++ src/lib.rs
@@ -108,9 +108,9 @@
 //! ## Traversing an `Element`
 //!
 //! ```
-//! # use ion_rs::IonResult;
+//! use ion_rs::IonResult;
 //! # fn main() -> IonResult<()> {
-//! use ion_rs::element::{Element, IntoAnnotatedElement, IonSequence, Value};
+//! use ion_rs::element::{Element, IonSequence, IntoAnnotatedElement,  Value};
 //! use ion_rs::{ion_struct, ion_list};
 //! let element: Element = ion_struct! {
 //!   "foo": "hello",
@@ -135,9 +135,9 @@
 //! ## Writing an `Element` to an `io::Write`
 //!
 //! ```
-//! # use ion_rs::IonResult;
+//! use ion_rs::IonResult;
 //! # fn main() -> IonResult<()> {
-//! use ion_rs::element::{Element, IntoAnnotatedElement, IonSequence, Value};
+//! use ion_rs::element::{Element, IonSequence, IntoAnnotatedElement,  Value};
 //! use ion_rs::{ion_struct, ion_list, TextWriterBuilder, IonWriter};
 //! use ion_rs::element::writer::ElementWriter;
 //! let element: Element = ion_struct! {
